% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-commands.R
\name{docker_env}
\alias{docker_env}
\title{Set an environment variable in the image build}
\usage{
docker_env(dockerfile = new_dockerfile(), ..., description = NULL)
}
\arguments{
\item{dockerfile}{Object of class "dockerfile". Create a new Dockerfile with
\code{\link[=new_dockerfile]{new_dockerfile()}} or \code{\link[=from_rocker]{from_rocker()}}.}

\item{...}{Named environment values to declare in the Docker container. These
will be evaluated immediately.}

\item{description}{Character. Use this description for all environment
variables. This only affects the Dockerfile description, and has no impact
on the actual Dockerfile.}
}
\value{
Object of class "dockerfile"
}
\description{
Set an environment variable in the image build
}
