% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-commands.R
\name{docker_run}
\alias{docker_run}
\title{RUN a custom command in image build}
\usage{
docker_run(dockerfile = new_dockerfile(), ..., description = NULL)
}
\arguments{
\item{dockerfile}{Object of class "dockerfile". Create a new Dockerfile with
\code{\link[=new_dockerfile]{new_dockerfile()}} or \code{\link[=from_rocker]{from_rocker()}}.}

\item{...}{Optionally named commands to \code{RUN}. Each command will be executed
separately (see Description). Command names are used to label commands in
the Dockerfile description, but have no impact on the actual Dockerfile.}

\item{description}{Character. Use this description for all commands. This
only affects the Dockerfile description, and has no impact on the actual
Dockerfile.}
}
\value{
Object of class "dockerfile"
}
\description{
Information is not shared between \code{RUN} commands. This means that an
environment variable set in one \code{RUN} command will not be available in
another. File system changes will be persisted, however. For example, if a
\code{RUN} command creates a file, that file will be accessible by other \code{RUN}
commands.
}
\examples{
dockerfile <- new_dockerfile()
docker_run(dockerfile, 'echo "Giraffes are tall"')
docker_run(
  dockerfile,
  echo_giraffe = 'echo "Giraffes are tall"',
  list_files = 'ls'
)

}
