% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent-images.R
\name{from_rocker}
\alias{from_rocker}
\title{Use rocker as a starting point for a new Dockerfile}
\usage{
from_rocker(
  base = c("r-base", "r-devel", "r-ver", "shiny", "rstudio", "tidyverse", "verse",
    "geospatial", "drd", "r-devel-san", "r-devel-ubsan-clang", "rstudio:testing",
    "r-apt"),
  version = "latest"
)
}
\arguments{
\item{version}{Character.}
}
\value{

}
\description{
Use rocker as a starting point for a new Dockerfile
}
