% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-commands.R
\name{install_r}
\alias{install_r}
\title{Install R}
\usage{
install_r(dockerfile, version = "latest")
}
\arguments{
\item{version}{}
}
\value{

}
\description{
This command should not be used with Rocker parent images, as these already
have R installed.
}
