% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-commands.R
\name{install_r_packages}
\alias{install_r_packages}
\title{Install R packages in a Docker image}
\usage{
install_r_packages(dockerfile, ..., CRAN = NULL, description = NULL)
}
\arguments{
\item{dockerfile}{Object of class "dockerfile".}

\item{...}{Unquoted package names.}

\item{CRAN}{Character. Optionally specify a CRAN repository from which to
install}
}
\value{

}
\description{
Use a dockerfile that installs R before attempting to install R packages.
This can be done by \verb{[install_r()]}, or by using a dockerfile that started with
\verb{[from_rocker()]}.
}
