% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{print_commands}
\alias{print_commands}
\title{Print Docker commands as they would appear in a Dockerfile}
\usage{
print_commands(dockerfile)
}
\arguments{
\item{dockerfile}{}
}
\value{

}
\description{
Print Docker commands as they would appear in a Dockerfile
}
